adding bintray as publisher
