+++ project/Environment.scala
@@ -22,6 +22,7 @@ object Environment {
 
   lazy val hadoopVersion  = either("SPARK_HADOOP_VERSION", "2.2.0")
   lazy val sparkVersion   = either("SPARK_VERSION", "1.3.1")
+  lazy val versionSuffix  = either("GEOTRELLIS_VERSION_SUFFIX", "-SNAPSHOT")
 
   lazy val javaGdalDir    = either("JAVA_GDAL_DIR", "/usr/local/lib")
 }
