+++ project/Info.scala
@@ -1,0 +1,6 @@
+object Info {
+  val description =
+    "GeoTrellis is an open source geographic data processing engine for high performance applications."
+  val url = "http://geotrellis.github.io"
+  val tags = Seq("maps", "gis", "geographic", "data", "raster", "processing")
+}
