+++ project/plugins.sbt
@@ -1,20 +1,16 @@
-credentials += Credentials(Path.userHome / ".ivy2" / ".credentials")
-
-addSbtPlugin("me.lessis" % "ls-sbt" % "0.1.3")
-
 resolvers ++= Seq(
   Classpaths.sbtPluginReleases,
   Opts.resolver.sonatypeReleases
 )
 
-addSbtPlugin("io.spray" % "sbt-revolver" % "0.7.2")
-
 addSbtPlugin("net.virtual-void" % "sbt-dependency-graph" % "0.7.5")
 
 addSbtPlugin("com.typesafe.sbt" % "sbt-pgp" % "0.8.2")
 
-addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.11.2")
+addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.13.0")
 
 addSbtPlugin("com.eed3si9n" % "sbt-dirty-money" % "0.1.0")
 
 addSbtPlugin("me.lessis" % "ls-sbt" % "0.1.3")
+
+addSbtPlugin("me.lessis" % "bintray-sbt" % "0.3.0")
